+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/web/PurchaseOrderLineController.java
@@ -17,6 +17,9 @@
  */
 package com.axelor.apps.purchase.web;
 
+import java.math.BigDecimal;
+import java.util.Map;
+
 import com.axelor.apps.account.db.Tax;
 import com.axelor.apps.account.db.TaxEquiv;
 import com.axelor.apps.account.db.TaxLine;
@@ -27,18 +30,15 @@ import com.axelor.apps.purchase.db.PurchaseOrder;
 import com.axelor.apps.purchase.db.PurchaseOrderLine;
 import com.axelor.apps.purchase.exception.IExceptionMessage;
 import com.axelor.apps.purchase.service.PurchaseOrderLineService;
-import com.axelor.apps.purchase.service.PurchaseOrderLineServiceImpl;
 import com.axelor.db.mapper.Mapper;
 import com.axelor.exception.AxelorException;
+import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
 import com.axelor.rpc.Context;
 import com.google.inject.Inject;
-
-import java.math.BigDecimal;
-import java.util.Map;
 
 public class PurchaseOrderLineController {
 
@@ -48,52 +48,17 @@ public class PurchaseOrderLineController {
 	public void compute(ActionRequest request, ActionResponse response) throws AxelorException{
 
 		Context context = request.getContext();
-		
 		PurchaseOrderLine purchaseOrderLine = context.asType(PurchaseOrderLine.class);
-
 		PurchaseOrder purchaseOrder = this.getPurchaseOrder(context);
-		
-		Product product = purchaseOrderLine.getProduct();
 
 		try{
-			if(purchaseOrder == null || purchaseOrderLine.getPrice() == null || purchaseOrderLine.getQty() == null)  {  return;  }
-
-			BigDecimal exTaxTotal = BigDecimal.ZERO;
-			BigDecimal companyExTaxTotal = BigDecimal.ZERO;
-			BigDecimal inTaxTotal = BigDecimal.ZERO;
-			BigDecimal companyInTaxTotal = BigDecimal.ZERO;
-			BigDecimal priceDiscounted = purchaseOrderLineService.computeDiscount(purchaseOrderLine);
-			response.setValue("priceDiscounted", priceDiscounted);
-			response.setAttr("priceDiscounted", "hidden", priceDiscounted.compareTo(purchaseOrderLine.getPrice()) == 0);
-			
-			BigDecimal taxRate = BigDecimal.ZERO;
-			if(purchaseOrderLine.getTaxLine() != null)  {  taxRate = purchaseOrderLine.getTaxLine().getValue();  }
-			
-			if(!purchaseOrder.getInAti()){
-				exTaxTotal = PurchaseOrderLineServiceImpl.computeAmount(purchaseOrderLine.getQty(), purchaseOrderLineService.computeDiscount(purchaseOrderLine));
-				inTaxTotal = exTaxTotal.add(exTaxTotal.multiply(taxRate));
-				companyExTaxTotal = purchaseOrderLineService.getCompanyExTaxTotal(exTaxTotal, purchaseOrder);
-				companyInTaxTotal = companyExTaxTotal.add(companyExTaxTotal.multiply(taxRate));
-			}
-			else  {
-				inTaxTotal = PurchaseOrderLineServiceImpl.computeAmount(purchaseOrderLine.getQty(), purchaseOrderLineService.computeDiscount(purchaseOrderLine));
-				exTaxTotal = inTaxTotal.divide(taxRate.add(BigDecimal.ONE), 2, BigDecimal.ROUND_HALF_UP);
-				companyInTaxTotal = purchaseOrderLineService.getCompanyExTaxTotal(inTaxTotal, purchaseOrder);
-				companyExTaxTotal = companyInTaxTotal.divide(taxRate.add(BigDecimal.ONE), 2, BigDecimal.ROUND_HALF_UP);
-			}
-
-			if (product != null) {
-				response.setValue("saleMinPrice", purchaseOrderLineService.getMinSalePrice(purchaseOrder, purchaseOrderLine));
-				response.setValue("salePrice", purchaseOrderLineService.getSalePrice(purchaseOrder, product, purchaseOrderLine.getPrice()));
-			}
-			response.setValue("exTaxTotal", exTaxTotal);
-			response.setValue("inTaxTotal", inTaxTotal);
-			response.setValue("companyExTaxTotal", companyExTaxTotal);
-			response.setValue("companyInTaxTotal", companyInTaxTotal);
+			Map<String, Object> map = purchaseOrderLineService.compute(purchaseOrderLine, purchaseOrder);
+			response.setValues(map);
+			response.setAttr("priceDiscounted", "hidden", ((BigDecimal) map.getOrDefault("priceDiscounted", BigDecimal.ZERO)).compareTo(purchaseOrderLine.getPrice()) == 0);
 
 		}
 		catch(Exception e)  {
-			response.setFlash(e.getMessage());
+			TraceBackService.trace(response, e);
 		}
 	}